+++ src/components/RegistrationTable/TableFooter.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/components/ui/button"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { DropdownMenu, DropdownMenuTrigger, DropdownMenuContent, DropdownMenuItem } from "@/components/ui/dropdown-menu"
 import { Table } from "@tanstack/react-table"
-import { exportToCSV, exportToPDF } from "@/util/exportHelpers"
+import { exportToCSV, exportToPDF, formatTableForExport } from "@/util/exportHelpers"
 
 import DownloadIcon from "../../../public/assets/icons/download_icon.svg"
 import { Attendee } from "./columns"
@@ -15,32 +15,14 @@ interface TableFooterProps {
     setPageSize: (size: number) => void
 }
 
-function createAttendeeTemplate(keys: string[]): Attendee {
-    return keys.reduce((template, key) => {
-        template[key] = null;
-        return template;
-    }, {} as Attendee);
-}
-
 export const TableFooter: React.FC<TableFooterProps> = ({ table, pageSize, setPageSize }) => {
+    const rows = formatTableForExport(table);
     const handleExport = (format: string) => {
-        const csvCols = [];
-        const csvRows: Attendee[] = [];
-        const headerCols = table.getRowModel().rows[0].getVisibleCells();
-        for (let i = 2; i < headerCols.length; i++) {
-            csvCols.push(headerCols[i].column.id);
+        if (format === "pdf") {
+            exportToPDF(rows);
+        } else {
+            exportToCSV(rows);
         }
-        const tableRows = Object.values(table.getCoreRowModel().rowsById);
-        for (const row of tableRows) {
-            const rowData = row.getVisibleCells();
-            const attendeeTemplate: Attendee = createAttendeeTemplate(csvCols);
-            for (let i = 2; i < rowData.length; i++) {
-                attendeeTemplate[rowData[i].column.id] = rowData[i].getValue();
-            }
-            csvRows.push(attendeeTemplate);
-        }
-        // exportToCSV(csvRows);
-        exportToPDF(csvRows);
         console.log(`Exporting as ${format}... (STUB)`);
     }
